New rTorrent/LibTorrent/ruTorrent and fixed for Debian (mostly)
Updated to rTorrent 0.9.8, LibTorrent 0.13.8, and ruTorrent 3.8 along with other minor software updates. Fixed issues with autodl-irssi and rtorrent not starting. Tested and confirmed working with lighttpd on Ubuntu 16.04 LTS, Ubuntu 18.04 LTS, Debian 8.11, Debian 9.9, and Debian 10.0. Nginx and Apache work on Ubuntu 18.04 LTS and do NOT work on Debian 10. Nginx and Apache were not tested on any older versions. Work will continue to get Nginx and Apache working on Debian 10, but for now they are broken.